Adding hungarian matching algorithm and tested cluster object re-assocaition

+++ dynosam/include/dynosam/dataprovider/DatasetProvider.hpp
@@ -297,13 +297,13 @@ public:
     }
 
 
-    virtual size_t datasetSize() const override {
+    virtual int datasetSize() const override {
         //this is very gross - the ending_frame_id_ will only get set after the first call to spin, but this is what we want to return
         //so the value may change depending on if setEndingFrame has been called and/or when datasetSize is called!!!!!
         if(is_first_spin_) {
-            return BaseDynoDataset::getDatasetSize();
+            return (int)BaseDynoDataset::getDatasetSize();
         }
-        return ending_frame_id_;
+        return (int)ending_frame_id_;
     }
 
     virtual bool spin() override {
